Chore: move main args match from main.rs to cli_parse module to handle_cli
